:hammer: refactor: GameElement 클래스 정의 cpp 파일로 분리
