+++ srcs/Server/includes/AGameElement.hpp
@@ -1,6 +1,8 @@
 #ifndef AGAMEELEMENT_HPP
 #define AGAMEELEMENT_HPP
 
+#pragma once
+
 class AGameElement
 {
 private:
@@ -17,26 +19,5 @@ public:
     virtual ~AGameElement();
     virtual void move() = 0; // 순수 가상 함수
 };
-
-AGameElement::AGameElement(float x, float y, float speed) : _x(x), _y(y), _speed(speed)
-{
-}
-
-AGameElement::AGameElement(const AGameElement &other) : _x(other._x), _y(other._y), _speed(other._speed)
-{
-}
-
-AGameElement &AGameElement::operator=(const AGameElement &other)
-{
-    if (this != &other)
-    {
-        this->_x = other._x;
-        this->_y = other._y;
-        this->_speed = other._speed;
-    }
-    return *this;
-}
-
-AGameElement::~AGameElement() {}
 
 #endif
