+++ srcs/Server/includes/Paddle.hpp
@@ -1,6 +1,8 @@
 #ifndef PADDLE_HPP
 #define PADDLE_HPP
 
+#pragma once
+
 #include "AGameElement.hpp"
 #include <algorithm>
 
@@ -34,55 +36,17 @@ public:
     Paddle &operator=(const Paddle &other);
     ~Paddle();
 
-    int &getLife() { return _life; }
-    float getX() const { return _x; }
-    float getY() const { return _y; }
-    float getSpeed() const { return _speed; }
-    float getWidth() const { return _width; }
-    float getHeight() const { return _height; }
-    e_paddle getDirection() const { return _direction; }
+    int &getLife();
+    float getX() const;
+    float getY() const;
+    float getSpeed() const;
+    float getWidth() const;
+    float getHeight() const;
+    e_paddle getDirection() const;
 
-    void move() override
-    {
-        if (this->_direction == STOP)
-            return;
-        else if (this->_direction == UP)
-            this->_y = std::max(this->_y + this->_speed, 1.0f - this->_height);
-        else if (this->_direction == DOWN)
-            this->_y = std::min(this->_y - this->_speed, -1.0f);
-    }
-    void setDirection(e_paddle direction)
-    {
-        this->_direction = direction;
-    }
+    void move() override;
+    void setDirection(e_paddle direction);
 };
 
-Paddle::Paddle(e_paddle player_num) : AGameElement(0.0f, 0.0f, PADDLE_SPEED), _width(PADDLE_WIDTH), _height(PADDLE_HEIGHT), _direction(STOP), _life(5)
-{
-    if (player_num == PLAYER1)
-        this->_x = -1.0f;
-    else if (player_num == PLAYER2)
-        this->_x = 1.0f - this->_width;
-}
-
-Paddle::Paddle(const Paddle &other) : AGameElement(other._x, other._y, other._speed), _width(other._width), _height(other._height), _direction(other._direction)
-{
-}
-
-Paddle &Paddle::operator=(const Paddle &other)
-{
-    if (this != &other)
-    {
-        this->_x = other._x;
-        this->_y = other._y;
-        this->_speed = other._speed;
-        this->_width = other._width;
-        this->_height = other._height;
-        this->_direction = other._direction;
-    }
-    return *this;
-}
-
-Paddle::~Paddle() {}
 
 #endif
